requirements.txt
anyio==3.4.0
asgiref==3.4.1
click==8.0.3
colorama==0.4.4
fastapi==0.70.1
h11==0.12.0
httptools==0.3.0
idna==3.3
numpy==1.21.5
opencv-python==4.5.4.60
Pillow==8.4.0
pydantic==1.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
typing_extensions==4.0.1
uvicorn==0.16.0
watchgod==0.7
websockets==10.1
